+++ src/roq/okx/gateway/drop_copy.cpp
@@ -145,6 +145,7 @@ DropCopy::DropCopy(Handler &handler, io::Context &context, uint16_t stream_id, A
           .error = create_metrics(shared.settings, name_, "error"sv),
           .subscribe = create_metrics(shared.settings, name_, "subscribe"sv),
           .unsubscribe = create_metrics(shared.settings, name_, "unsubscribe"sv),
+          .notice = create_metrics(shared.settings, name_, "notice"sv),
           .channel_conn_count = create_metrics(shared.settings, name_, "channel_conn_count"sv),
           .login = create_metrics(shared.settings, name_, "login"sv),
           .account = create_metrics(shared.settings, name_, "account"sv),
@@ -194,6 +195,7 @@ void DropCopy::operator()(metrics::Writer &writer) const {
       .write(profile_.error, metrics::Type::PROFILE)
       .write(profile_.subscribe, metrics::Type::PROFILE)
       .write(profile_.unsubscribe, metrics::Type::PROFILE)
+      .write(profile_.notice, metrics::Type::PROFILE)
       .write(profile_.channel_conn_count, metrics::Type::PROFILE)
       .write(profile_.login, metrics::Type::PROFILE)
       .write(profile_.account, metrics::Type::PROFILE)
@@ -467,6 +469,13 @@ void DropCopy::operator()(Trace<protocol::json::Unsubscribe> const &event) {
   profile_.unsubscribe([&]() {
     auto &[trace_info, unsubscribe] = event;
     log::info<1>("unsubscribe={}"sv, unsubscribe);
+  });
+}
+
+void DropCopy::operator()(Trace<protocol::json::Notice> const &event) {
+  profile_.notice([&]() {
+    auto &[trace_info, notice] = event;
+    log::warn("notice={}"sv, notice);
   });
 }