+++ src/roq/okx/gateway/market_data.cpp
@@ -86,6 +86,7 @@ MarketData::MarketData(Handler &handler, io::Context &context, uint16_t stream_i
           .error = create_metrics(shared.settings, name_, "error"sv),
           .subscribe = create_metrics(shared.settings, name_, "subscribe"sv),
           .unsubscribe = create_metrics(shared.settings, name_, "unsubscribe"sv),
+          .notice = create_metrics(shared.settings, name_, "notice"sv),
           .login = create_metrics(shared.settings, name_, "login"sv),
           .estimated_price = create_metrics(shared.settings, name_, "estimated_price"sv),
           .price_limit = create_metrics(shared.settings, name_, "price_limit"sv),
@@ -129,6 +130,7 @@ void MarketData::operator()(metrics::Writer &writer) const {
       .write(profile_.error, metrics::Type::PROFILE)
       .write(profile_.subscribe, metrics::Type::PROFILE)
       .write(profile_.unsubscribe, metrics::Type::PROFILE)
+      .write(profile_.notice, metrics::Type::PROFILE)
       .write(profile_.login, metrics::Type::PROFILE)
       .write(profile_.estimated_price, metrics::Type::PROFILE)
       .write(profile_.price_limit, metrics::Type::PROFILE)
@@ -375,6 +377,13 @@ void MarketData::operator()(Trace<protocol::json::Unsubscribe> const &event) {
   profile_.unsubscribe([&]() {
     auto &[trace_info, unsubscribe] = event;
     log::info<1>("unsubscribe={}"sv, unsubscribe);
+  });
+}
+
+void MarketData::operator()(Trace<protocol::json::Notice> const &event) {
+  profile_.notice([&]() {
+    auto &[trace_info, notice] = event;
+    log::warn("notice={}"sv, notice);
   });
 }