+++ src/roq/okx/gateway/market_data.hpp
@@ -85,6 +85,7 @@ struct MarketData final : public web::socket::Client::Handler, public protocol::
   void operator()(Trace<protocol::json::Error> const &) override;
   void operator()(Trace<protocol::json::Subscribe> const &) override;
   void operator()(Trace<protocol::json::Unsubscribe> const &) override;
+  void operator()(Trace<protocol::json::Notice> const &) override;
 
   void operator()(Trace<protocol::json::Status> const &) override;
   void operator()(Trace<protocol::json::Instruments> const &) override;
@@ -129,8 +130,8 @@ struct MarketData final : public web::socket::Client::Handler, public protocol::
     utils::metrics::Counter disconnect;
   } counter_;
   struct {
-    utils::metrics::Profile parse, error, subscribe, unsubscribe, login, estimated_price, price_limit, mark_price, tickers, trades, bbo_tbt, books_l2_tbt,
-        index_tickers, funding_rate;
+    utils::metrics::Profile parse, error, subscribe, unsubscribe, notice, login, estimated_price, price_limit, mark_price, tickers, trades, bbo_tbt,
+        books_l2_tbt, index_tickers, funding_rate;
   } profile_;
   struct {
     utils::metrics::Latency ping, heartbeat;