+++ src/roq/okx/gateway/static_data.hpp
@@ -90,6 +90,7 @@ struct StaticData final : public web::socket::Client::Handler, public protocol::
   void operator()(Trace<protocol::json::Error> const &) override;
   void operator()(Trace<protocol::json::Subscribe> const &) override;
   void operator()(Trace<protocol::json::Unsubscribe> const &) override;
+  void operator()(Trace<protocol::json::Notice> const &) override;
 
   void operator()(Trace<protocol::json::Status> const &) override;
   void operator()(Trace<protocol::json::Instruments> const &) override;
@@ -133,7 +134,7 @@ struct StaticData final : public web::socket::Client::Handler, public protocol::
     utils::metrics::Counter disconnect;
   } counter_;
   struct {
-    utils::metrics::Profile parse, error, subscribe, unsubscribe, login, status, instruments;
+    utils::metrics::Profile parse, error, subscribe, unsubscribe, notice, login, status, instruments;
   } profile_;
   struct {
     utils::metrics::Latency ping, heartbeat;